src/requirements.txt
fastapi
pydantic
pydantic[email]
pydantic_settings
starlette
uvicorn
python-dotenv
mysql-connector-python